LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.radial.cone.front-expected.txt

2d.gradient.radial.cone.front


FAIL OffscreenCanvas test: 2d.gradient.radial.cone.front assert_equals: Green channel of the pixel at (98, 25) expected 255 but got 254
